Negamax and Cleaning
To reflect the modular structure of the decision driver, classifier,
heuristic AIs IDecisionDriver was created and the separate classes were
split into MCUTC, Minimax, and Negamax. The alpha-beta pruning drivers
should really derive from the same abstract class, but that will be done
when NegaScout is introduced. Decision tree and game definitions have
been untouched, but the actual program running was modified to facility
AI testing.